requirements.txt
grammar_graph>=0.1.0
proxyorderedset>=0.3.0
pydot>=1.4.2
wheel>=0.37.1
z3-solver>=4.8.17.0
